Remove unnecessary padding from texture array constructor
Originally this padding was thought necessary as it fixed an issue where
the texture would gradually become more vertically offset as each frame
progressed. However, it appears that this fix is only necessary for
Linux+Vulkan, as Windows and macOS both behave correctly without the
added row padding and adding the padding results in crashing or
incorrectly offset textures respectively.

The docs **do** say that that the extra padding requirements should not
apply to the `Queue::write_texture` method:

https://docs.rs/wgpu/0.10.1/wgpu/struct.ImageDataLayout.html#structfield.bytes_per_row